Minor refactor of plugin tests
